requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Deep Learning
tensorflow>=2.8.0
torch>=1.11.0
pytorch-lightning>=1.6.0

# Anomaly Detection Libraries
pyod>=1.0.0
adtk>=0.6.2
stumpy>=1.11.0

# Time Series
tslearn>=0.5.2
tsfresh>=0.19.0
sktime>=0.11.0

# Configuration and CLI
PyYAML>=6.0
click>=8.0.0
rich>=12.0.0
colorama>=0.4.4
tqdm>=4.62.0

# Data Processing
joblib>=1.1.0
h5py>=3.6.0
tables>=3.7.0

# Evaluation and Metrics
imbalanced-learn>=0.8.0
yellowbrick>=1.4.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
flake8>=4.0.0
black>=22.0.0

# Optional: Advanced features
optuna>=2.10.0
hyperopt>=0.2.7